pyproject.toml
[tool.poetry]
name = "jarvis-azure"
version = "0.1.0"
description = "Jarvis AI bot, powered by Azure OpenAI"
authors = ["Yusuf Kaka <yusufk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.26.1"
python-telegram-bot = {extras = ["webhooks"], version = "^20.0"}
python-dotenv = "^0.21.1"
langchain = "^0.0.152"
tiktoken = "^0.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
